Split specific client stitch uses into different modules
