+++ src/utils/snooty-stitch.js
@@ -1,0 +1,13 @@
+import { SNOOTY_STITCH_ID } from '../build-constants';
+import { callStitchFunction } from './stitch-common';
+
+export const callSnootyStitchFunction = async (fnName, metadata, ...fnArgs) => {
+    try {
+        return callStitchFunction(fnName, SNOOTY_STITCH_ID, [
+            metadata,
+            ...fnArgs,
+        ]);
+    } catch (error) {
+        console.error(error);
+    }
+};
